+++ Scripts/Player/DefineStatut.cs
@@ -7,7 +7,7 @@ public class DefineStatut : MonoBehaviour
     public float countS;
     public float coefSpeed;
     public float countSW;
-    public int countD;
-    public int hp;
-    public int armor;
+    public float countD;
+    public float hp;
+    public float armor;
 }
